requirements.txt
flask~=2.0.1
flask_login
flask_sqlalchemy
flask_restx
werkzeug~=2.0.2
pycrypto~=1.4.1
setuptools~=59.2.0
cryptography~=3.4.7
pandas~=1.3.1